First attempt at polymorphic old expression

+++ test/solc-verify/OldExpr.sol
@@ -2,24 +2,24 @@ pragma solidity >=0.5.0;
 
 contract OldExpr {
     address owner;
-    int256 counter;
+    int8 counter;
 
     constructor() public {
         owner = msg.sender;
     }
 
-    /** @notice postcondition msg.sender == __verifier_old_address(owner) || owner == __verifier_old_address(owner) */
+    /** @notice postcondition msg.sender == __verifier_old(owner) || owner == __verifier_old(owner) */
     function changeOwnerCorrect() public {
         require(msg.sender == owner);
         owner = msg.sender;
     }
 
-    /** @notice postcondition msg.sender == __verifier_old_address(owner) || owner == __verifier_old_address(owner) */
+    /** @notice postcondition msg.sender == __verifier_old(owner) || owner == __verifier_old(owner) */
     function changeOwnerIncorrect() public {
         owner = msg.sender;
     }
 
-    /** @notice postcondition counter == __verifier_old_int256(counter) + 1 */
+    /** @notice postcondition counter == __verifier_old(counter) + 1 */
     function increment() public {
         counter++;
     }
